requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
keras==2.15.0
ml-dtypes==0.2.0
mediapipe==0.10.14
opencv-python==4.9.0.80
pandas==2.2.2
numpy==1.24.3
scikit-learn==1.5.0
openai==1.30.1
python-dotenv==1.0.1